Refactor teachers app
