+++ src/apps/teachers/api/serializers/__init__.py
@@ -1,0 +1,2 @@
+from .teacher_serializers import TeacherViewSerializer, TeacherCreationSerializer
+from .schedule_serializers import ScheduleViewSerializer, ScheduleCreationSerializer
