+++ src/apps/teachers/api/views/__init__.py
@@ -1,0 +1,2 @@
+from .teacher_view_set import TeacherViewSet
+from .schedule_view_set import ScheduleViewSet
