+++ src/apps/teachers/models/__init__.py
@@ -1,0 +1,2 @@
+from .teacher import Teacher
+from .schedule import Schedule
